+++ src/main/java/com/bennie/spring/service/PersonServiceImpl.java
@@ -22,7 +22,7 @@ public class PersonServiceImpl implements PersonService {
 	}
 
 	@Override
-	public Person get(Long id) {
+	public Person get(Integer id) {
 
 		for (Person person : persons) {
 			if (person.getId().longValue() == id.longValue())
@@ -34,7 +34,7 @@ public class PersonServiceImpl implements PersonService {
 
 	@Override
 	public Person add(Person person) {
-		Long newId = 0L;
+		Integer newId = 0;
 		Boolean hasFoundSuitableId = false;
 		while (hasFoundSuitableId == false) {
 			for (int i = 0; i < persons.size(); i++) {
@@ -56,7 +56,7 @@ public class PersonServiceImpl implements PersonService {
 	}
 
 	@Override
-	public Boolean delete(Long id) {
+	public Boolean delete(Integer id) {
 		for (Person person : persons) {
 			if (person.getId().longValue() == id.longValue()) {
 				persons.remove(person);
@@ -87,19 +87,19 @@ public class PersonServiceImpl implements PersonService {
 
 	private void populateData() {
 		Person person1 = new Person();
-		person1.setId(0L);
+		person1.setId(0);
 		person1.setFirstName("John");
 		person1.setLastName("Smith");
 		person1.setMoney(1000.0);
 
 		Person person2 = new Person();
-		person2.setId(1L);
+		person2.setId(1);
 		person2.setFirstName("Jane");
 		person2.setLastName("Adams");
 		person2.setMoney(2000.0);
 
 		Person person3 = new Person();
-		person3.setId(2L);
+		person3.setId(2);
 		person3.setFirstName("Jeff");
 		person3.setLastName("Mayer");
 		person3.setMoney(3000.0);
